requirements.txt
# PDF/PowerPoint処理 (macOS最適化)
Pillow==10.2.0
python-pptx==0.6.23
PyMuPDF==1.23.26

# Google Gemini API
google-generativeai==0.3.2

# Excel/データ出力
openpyxl==3.1.2
pandas==2.1.4

# 実行ファイル化 (macOS)
pyinstaller==6.3.0

# ユーティリティ
python-dotenv==1.0.0
tqdm==4.66.1
pathvalidate==3.2.0
numpy==1.26.3
pyyaml==6.0.1

# GUI対応は標準のtkinterを使用

# macOS特有
pyobjc-framework-Cocoa==10.1; sys_platform == "darwin"
pyobjc-framework-Quartz==10.1; sys_platform == "darwin"